refactor: firebase/lite substitution / navigation bar component splitting/ bundle-visualisation implement
