+++ my-monorepo/apps/shopease/package.json
@@ -61,6 +61,7 @@
     "eslint-plugin-storybook": "^0.10.1",
     "globals": "^15.9.0",
     "postcss": "^8.4.47",
+    "rollup-plugin-visualizer": "^5.12.0",
     "tailwindcss": "^3.4.14",
     "typescript": "^5.5.3",
     "typescript-eslint": "^8.7.0",
