+++ my-monorepo/apps/shopease/vite.config.ts
@@ -1,6 +1,7 @@
 //vite.config.ts
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
+import { visualizer } from "rollup-plugin-visualizer";
 import path from "path";
 import { fileURLToPath } from "url";
 const __filename = fileURLToPath(import.meta.url);
@@ -8,10 +9,18 @@ const __dirname = path.dirname(__filename);
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()],
+  plugins: [
+    react(),
+    visualizer({ open: true, filename: "bundle-visualization.html" }),
+  ],
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "src/"),
+    },
+  },
+  build: {
+    rollupOptions: {
+      treeshake: true, // Ensure tree shaking is enabled
     },
   },
 });
